refactor: Button과 icons 폴더 구조 변경
