+++ src/components/common/icons/BankIcon.jsx
@@ -1,4 +1,4 @@
-import styles from './BankIcon.module.css';
+import styles from './css/BankIcon.module.css';
 import { IoMdBook } from 'react-icons/io';
 
 export default function BankIcon() {
