+++ src/components/common/icons/css/BankIcon.module.css
