+++ src/components/common/icons/css/MoneyIcon.module.css
