+++ src/components/common/icons/css/SetMenuIcon.module.css
