+++ src/components/common/icons/css/Symbole.module.css
